requirements.txt
pywebview
peewee
Flask
websockets
pyinstaller
psutil